+++ database/dbCommands.py
@@ -1,20 +1,21 @@
 import config.configHandler
 import config.dbConstants
 
-def findItem(itemName, database):
+
+def find_item(item_name, database):
     """Returns an item from the DB if it exists, else None"""
     for item in database:
-        if item.name == itemName:
+        if item.name == item_name:
             return item
     return None
 
 
-def addItem(item, database):
+def add_item(item, database):
     """Appends an item to the end of the database"""
     database.append(item)
 
 
-def updateItem(new_item, database):
+def update_item(new_item, database):
     """Updates an item to the end of the database"""
     for item in database:
         if item.name == new_item.name:
@@ -22,37 +23,10 @@ def updateItem(new_item, database):
             item.type = new_item.type
 
 
-class DbCommands:
-    ###########################
-    ## Database functions below
-    ###########################
-
-    def getPriceOfItem(self, name, quantity = 1):
-        """Returns the item's price multiplied by the optional quantity"""
-        for item in self.database:
-            if item[self.ProjectConfig.nameColumnIndex] == name:
-                if quantity == 1:
-                    return "Price of " + name + " is: $" + str(item[self.ProjectConfig.priceColumnIndex])
-                else:
-                    return "Price of  "+ str(quantity) + " " + name + " is: $" + str \
-                        (quantity * int(item[self.ProjectConfig.priceColumnIndex]))
-        return "Item: " + name + " not found!"
-
-    def addItemToDatabase(self, name, price, type = 'uncategorized'):
-        """Add the item to the database with the specified price and type"""
-        itemFound = False
-        for item in self.database:
-            # Try to find the item first, if it already exists
-            if item[self.ProjectConfig.nameColumnIndex] == name:
-                itemFound = True
-                item[self.ProjectConfig.priceColumnIndex] = price
-                item[self.ProjectConfig.typeColumnIndex] = type
-                return "Item: " + name + " updated!"
-
-        if not itemFound:
-            newItem = [1 ,2 ,3]
-            newItem[self.ProjectConfig.nameColumnIndex] = name
-            newItem[self.ProjectConfig.priceColumnIndex] = price
-            newItem[self.ProjectConfig.typeColumnIndex] = type
-            self.database.append(newItem)
-            return "Item: " + str(newItem) + " added to database!"
+def delete_item(item_to_delete_name, database):
+    """Attempt to delete an item. Return true if successful"""
+    for item in database:
+        if item.name == item_to_delete_name:
+            database.remove(item)
+            return True
+    return False